Refactor: move RaftMsg and ExternalCommand to separate file
