+++ openraft/src/core/raft_core.rs
@@ -27,6 +27,13 @@ use crate::config::RuntimeConfig;
 use crate::core::balancer::Balancer;
 use crate::core::command_state::CommandState;
 use crate::core::notify::Notify;
+use crate::core::raft_msg::external_command::ExternalCommand;
+use crate::core::raft_msg::AppendEntriesTx;
+use crate::core::raft_msg::ClientWriteTx;
+use crate::core::raft_msg::InstallSnapshotTx;
+use crate::core::raft_msg::RaftMsg;
+use crate::core::raft_msg::ResultSender;
+use crate::core::raft_msg::VoteTx;
 use crate::core::sm;
 use crate::core::sm::CommandSeq;
 use crate::core::ServerState;
@@ -60,16 +67,9 @@ use crate::progress::Progress;
 use crate::quorum::QuorumSet;
 use crate::raft::AppendEntriesRequest;
 use crate::raft::AppendEntriesResponse;
-use crate::raft::AppendEntriesTx;
 use crate::raft::ClientWriteResponse;
-use crate::raft::ClientWriteTx;
-use crate::raft::ExternalCommand;
 use crate::raft::InstallSnapshotRequest;
-use crate::raft::InstallSnapshotTx;
-use crate::raft::RaftMsg;
-use crate::raft::ResultSender;
 use crate::raft::VoteRequest;
-use crate::raft::VoteTx;
 use crate::raft_state::LogStateReader;
 use crate::replication;
 use crate::replication::Replicate;
